requirements.txt
picovoice
pvrecorder
python-dotenv
